+++ src/main/java/study/posthub/global/config/SecurityConfig.java
@@ -1,14 +1,12 @@
 package study.posthub.global.config;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
-import study.posthub.domain.member.service.CustomOAuth2MemberService;
+import study.posthub.domain.member.service.impl.CustomOAuth2MemberService;
 
 @Configuration
 @EnableWebSecurity
